requirement.txt
openai
python-dotenv
opencv-python
flask
fpdf
requests